// src/app/_constants/enums.ts
export const enum SelectedTool {
    pencil = 1,
    spray = 2,
    square = 3, 
    circle = 4,
    polygon = 5,
    eraser = 6
}

export const enum AppType {
    browser = 0,
    paint = 1,
    messaging = 2,
    showimage = 3,
    video = 4,
    saveimage = 5,
    screensaver = 6,
    background = 7
}

export const enum CallType {
    answer = "answer",
    call = "call"
}

export const enum UserStatus {
    offline = "offline",
    online = "online"
}

export const enum SelectedScreensaver {
    pipes = "pipes",
    logos = "logos"
}

export const enum UserRole {
    systemAdmin = "system-admin",
    user = "user" 
}